[rs] Split `parser::tags` into `streaming::tag` and `complete::tag`
